feat: Add stereo image publisher node
This commit adds a new node called `stereo_image_publisher` to the `camera_delay` package. The node subscribes to the left and right camera image topics and combines the images into a stereo image. The combined stereo image is then published on the `/stereo_camera/image_raw` topic.

The `stereo_image_publisher` node uses the `cv_bridge` package to convert the ROS `Image` messages to OpenCV images. It then combines the left and right images side by side and publishes the stereo image.

This new node enhances the functionality of the `camera_delay` package by providing a convenient way to visualize the stereo images from the left and right cameras.